feat(router): refactor graphQL resolver
