+++ src/router/schema.ts
@@ -160,53 +160,3 @@ export async function usage(req, res) {
 		success: true,
 	});
 }
-
-export async function getSuperGraph(
-	req: Request,
-	res: Response,
-	next: NextFunction
-) {
-	const apolloServer = getServer();
-
-	try {
-		const { graphqlResponse, responseInit } = await runHttpQuery(
-			[],
-			{
-				method: req.method,
-				options: () =>
-					apolloServer.createGraphQLServerOptions(req, res),
-				query: {
-					query: `\
-query SuperGraphDl {
-  routerConfig {
-      __typename
-      id
-      supergraphSdl
-      minDelaySeconds
-  }
-}`,
-				},
-				request: convertNodeHttpToRequest(req),
-			},
-			apolloServer.csrfPreventionRequestHeaders
-		);
-
-		if (responseInit.headers) {
-			for (const [name, value] of Object.entries(responseInit.headers)) {
-				res.setHeader(name, value);
-			}
-		}
-		res.status(responseInit.status || 200).send(graphqlResponse);
-	} catch (error) {
-		if (!isHttpQueryError(error)) {
-			return next(error);
-		}
-
-		if (error.headers) {
-			for (const [name, value] of Object.entries(error.headers)) {
-				res.setHeader(name, value);
-			}
-		}
-		res.status(error.statusCode).send(error.message);
-	}
-}
