+++ test/integration/data/request/getRouterConfig.graphql
@@ -1,0 +1,6 @@
+query reportConfigQuery {
+	routerConfig {
+		id
+		supergraphSdl
+	}
+}
