+++ test/integration/data/response/getRouterConfig.json
@@ -1,0 +1,7 @@
+{
+  "data": {
+    "routerConfig": {
+      "id": "84c003"
+    }
+  }
+}
